# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "SPORTZONE_E-COMMERCE REST API built using Go, PSQL, REST API following Clean Architecture.",
        "title": "SPORTZONE_E-COMMERCE REST API",
        "contact": {},
        "version": "2.0"
    },
    "host": "localhost:3000",
    "paths": {
...
        "/admin/product": {
            "get": {
                "description": "LISTING ALL PRODUCTS FROM ADMINS AND USERS END",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ADMIN USER"
                ],
                "summary": "API FOR LISTING ALL PRODUCTS",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Enter the page number to display",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items to retrieve per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/admin/product/add": {
            "post": {
                "description": "ADDING PRODUCT FROM ADMINS END",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ADMIN"
                ],
                "summary": "API FOR ADDING PRODUCT",
                "operationId": "ADMIN-ADD-PRODUCT",
                "parameters": [
                    {
                        "description": "Enter the product details",
                        "name": "product_details",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/utils.AddProduct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/admin/product/bybrands/{brandid}": {
            "get": {
                "description": "LISTING ALL PRODUCTS FROM ADMINS AND USERS END BASED ON BRANDS",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ADMIN USER"
                ],
                "summary": "API FOR LISTING ALL PRODUCTS BASED ON CATEGBRANDSORY",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Enter the brand id",
                        "name": "brandid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Enter the page number to display",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items to retrieve per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/admin/product/bycategory/{categoryid}": {
            "get": {
                "description": "LISTING ALL PRODUCTS FROM ADMINS AND USERS END BASED ON CATEGORY",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ADMIN USER"
                ],
                "summary": "API FOR LISTING ALL PRODUCTS BASED ON CATEGORY",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Enter the category id",
                        "name": "categoryid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Enter the page number to display",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items to retrieve per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/admin/product/delete/{productid}": {
            "post": {
                "description": "DELETING PRODUCT BASED ON PRODUCT ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ADMIN"
                ],
                "summary": "API FOR DELETING A PRODUCT",
                "operationId": "ADMIN-DELETE-PRODUCT",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Enter the product id that you would like to delete",
                        "name": "productid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/admin/product/detail/add": {
            "post": {
                "description": "ADDING PRODUCT DETAILS FROM ADMINS END",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ADMIN"
                ],
                "summary": "API FOR ADDING PRODUCT DETAILS",
                "operationId": "ADMIN-ADD-PRODUCT-DETAILS",
                "parameters": [
                    {
                        "description": "Enter the product details",
                        "name": "product_details",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/utils.AddProductDetail"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/admin/product/detail/delete/{productdetailid}": {
            "delete": {
                "description": "DELETING PRODUCT DETAILS FROM ADMINS END",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ADMIN"
                ],
                "summary": "API FOR DELETE PRODUCT DETAILS",
                "operationId": "ADMIN-DELETE-PRODUCT-DETAILS",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Enter the product details id that you would like to delete",
                        "name": "productdetailid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/admin/product/detail/update/{productdetailid}": {
            "patch": {
                "description": "UPDATING PRODUCT DETAILS FROM ADMINS END",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ADMIN"
                ],
                "summary": "API FOR EDITING PRODUCT DETAILS",
                "operationId": "ADMIN-EDIT-PRODUCT-DETAILS",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Enter the product details id that you would like to make the change",
                        "name": "productdetailid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Enter the product details",
                        "name": "product_details",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/utils.AddProductDetail"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/admin/product/detail/{productid}": {
            "get": {
                "description": "LISTING ALL PRODUCTS DETAILS FROM ADMINS AND USERS END",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ADMIN USER"
                ],
                "summary": "API FOR LISTING PRODUCTS DETAILS BY ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Enter the product id that you would like to see the details of",
                        "name": "productid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Enter the page number to display",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items to retrieve",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
        "/user/filter/brands": {
            "get": {
                "description": "LISTING ALL BRANDS USERS END",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "USER"
                ],
                "summary": "API FOR LISTING ALL BRANDS",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "utils.AddProduct": {
            "type": "object",
            "required": [
                "categoryid",
                "image",
                "modelname"
            ],
            "properties": {
                "brandid": {
                    "type": "integer"
                },
                "categoryid": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "modelname": {
                    "type": "string"
                }
            }
        },
        "utils.AddProductDetail": {
            "type": "object",
            "required": [
                "availablecolourid",
                "availablesizeid",
                "price",
                "productid",
                "stock"
            ],
            "properties": {
                "availablecolourid": {
                    "type": "integer"
                },
                "availablesizeid": {
                    "type": "integer"
                },
                "discountid": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "productid": {
                    "type": "integer"
                },
                "stock": {
                    "type": "integer"
                }
            }
        },
...
        "utils.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        }
    }
}